many edits

+++ app/Controllers/website/WishListController.php
@@ -14,7 +14,8 @@ use \App\Models\Cart;
 class WishListController extends \App\Controllers\Controller{
   
     
- public function index($request,$response) {
+    // display all my wishlist Products
+    public function index($request,$response) {
        $wishlist = WishList::where('user_id',$_SESSION['auth-user'])->get()->toArray();
         $maincart = [];
         foreach($wishlist as $item ) {
@@ -24,46 +25,49 @@ class WishListController extends \App\Controllers\Controller{
             $mainwish[] =  array_merge($product,$item);
         }
    
-       return $this->container->view->render($response,'website/wish-list.twig',['wishlist'=>$mainwish]);
+       return $this->view->render($response,'website/wish-list.twig',['wishlist'=>$mainwish]);
     }
     
     
+    // add all the products to wish list
     public function alltocart($request,$response) {
-     
        $wishlist = WishList::where('user_id',$_SESSION['auth-user'])->get()->toArray();
-        
-        
        foreach($wishlist as $item ) {
             Cart::create([
                 'user_id' => $_SESSION['auth-user'],
                 'productID' => $item['productID'] ,
                 'quantity' => 1
             ]);
-            $this->flash->addMessage('success','تم إضافة جميع المنتجات الى السلة');
-            return $response->withStatus(302)->withHeader('Location', $this->router->urlFor('website.wishlist'));
-
+            $this->flashsuccess('تم إضافة جميع المنتجات الى السلة');
+           return $response->withRedirect($this->router->pathFor('website.wishlist'));
        }
-        
     }
     
+    
+    // add the product to the wishlist
     public function add($request,$response,$args) {
+        
+        // get the product id
         $id = rtrim($args['id'], '/');
-        WishList::create([
-            'user_id' => $_SESSION['auth-user'],
-            'productID' => $id ,
-        ]);
-        $this->flash->addMessage('success','تم إضافة المنتوج إلى قائمة الأمنيات بنجاح');
-        return $response->withStatus(302)->withHeader('Location', $this->router->urlFor('website.wishlist',['id'=>$id]));
+        
+        // add product to wish list
+        WishList::create([ 'user_id' => $_SESSION['auth-user'], 'productID' => $id ]);
+        
+        // flash success & redirect
+        $this->flashsuccess('تم إضافة المنتوج إلى قائمة الأمنيات بنجاح');
+        return $response->withRedirect($this->router->pathFor('website.wishlist',['id'=>$id]));
     }
     
     
-    
+    // delete the product from wishlist
     public function delete($request,$response,$args) {
         $id = rtrim($args['id'], '/');
         $wish = WishList::find($id);
-        $wish->delete();
-        $this->flash->addMessage('success','تم ازالة المنتوج من قائمة الأمنيات بنجاح');
-        return $response->withStatus(302)->withHeader('Location', $this->router->urlFor('website.product',['id'=>$id]));
+        if($wish){
+            $wish->delete();
+            $this->flashsuccess('تم ازالة المنتوج من قائمة الأمنيات بنجاح');
+        }
+        return $response->withRedirect($this->router->pathFor('website.product',['id'=>$id]));
     }
